15 - Finalizando o projeto e gravando o vídeo

+++ Projeto1Loja/Models/Produto.cs
@@ -6,6 +6,6 @@
         public string? nome { get; set; }
         public string? descricao { get; set; }
         public decimal preco { get; set; }
-        public decimal quantidade { get; set; }
+        public int quantidade { get; set; }
     }
 }
